+++ src/components/Navigation/Navbar/types.ts
@@ -12,6 +12,7 @@ export type NavbarNavigationItemsProps = {
 export type NavbarProps = {
   logo?: React.ReactNode
   linkRouter: any // Link from react router or next.js
+  pathname: string // Pathname from react router or next.js
   navigationSection?: NavbarNavigationItemsProps[]
   utilitySection?: React.ReactNode[]
   actionsSection?: {