+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - [IconMarkers] - from presets to Geospatial/IconMarker/presets/IconMarkers
 - [Pagination] - from DataDiplay to Navigation
 - [TableProps] - [TextareaProps] - [TextInputProps] - not required fields
+- [NavigationRail] - adding NavigationRailTabProps to README
 
 ## [v2.131.0] - 09-10-2025